backend/api/requirements.txt
# ==============================================================================
# backend/api/requirements.txt
# FastAPI service dependencies
# ==============================================================================

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Pydantic / Settings ────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1

# ── GCP SDKs ──────────────────────────────────────────────────────────────────
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.27.1
google-cloud-secret-manager==2.21.0
google-cloud-logging==3.11.3
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.74.0
vertexai==1.74.0

# ── Auth / Security ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Observability ─────────────────────────────────────────────────────────────
opentelemetry-sdk==1.28.4
opentelemetry-exporter-gcp-trace==1.8.0
opentelemetry-instrumentation-fastapi==0.49b4

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx==0.27.2
